+++ src/nock.utils.test.ts
@@ -46,6 +46,36 @@ describe('nock utils', () => {
         });
       });
 
+      it('removes apikey from raw headers', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          rawHeaders: ['api_key', 'secret'],
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'api-key',
+            type: SecurityType.APIKEY,
+            in: ApiKeyPlacement.HEADER,
+            name: 'api_key',
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          rawHeaders: ['api_key', HIDDEN_CREDENTIALS_PLACEHOLDER],
+        });
+      });
+
       it('removes apikey from body', () => {
         const definition: RecordingDefinition = {
           scope: 'https://localhost',
@@ -167,6 +197,48 @@ describe('nock utils', () => {
           status: 200,
         });
       });
+
+      it('removes apikey from response', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          body: {
+            my_api_key: 'secret',
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: 'secret' },
+          },
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'api-key',
+            type: SecurityType.APIKEY,
+            in: ApiKeyPlacement.BODY,
+            name: 'api_key',
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          body: {
+            my_api_key: HIDDEN_CREDENTIALS_PLACEHOLDER,
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: HIDDEN_CREDENTIALS_PLACEHOLDER },
+          },
+        });
+      });
     });
 
     describe('when removing basic auth credentials', () => {
@@ -203,6 +275,82 @@ describe('nock utils', () => {
           },
         });
       });
+
+      it('removes basic token from raw headers', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: 'Basic secret',
+          },
+          rawHeaders: ['Authorization', 'secret'],
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'basic',
+            type: SecurityType.HTTP,
+            scheme: HttpScheme.BASIC,
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: `Basic ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+          },
+          rawHeaders: ['Authorization', HIDDEN_CREDENTIALS_PLACEHOLDER],
+        });
+      });
+
+      it('removes basic token from response', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: 'Basic secret',
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: 'secret' },
+          },
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'basic',
+            type: SecurityType.HTTP,
+            scheme: HttpScheme.BASIC,
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: `Basic ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: HIDDEN_CREDENTIALS_PLACEHOLDER },
+          },
+        });
+      });
     });
 
     describe('when removing bearer token', () => {
@@ -236,6 +384,82 @@ describe('nock utils', () => {
           status: 200,
           reqheaders: {
             ['Authorization']: `Bearer ${TMP_PLACEHOLDER}`,
+          },
+        });
+      });
+
+      it('removes bearer token from raw headers', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: 'Bearer secret',
+          },
+          rawHeaders: ['Authorization', 'secret'],
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'bearer',
+            type: SecurityType.HTTP,
+            scheme: HttpScheme.BEARER,
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: `Bearer ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+          },
+          rawHeaders: ['Authorization', HIDDEN_CREDENTIALS_PLACEHOLDER],
+        });
+      });
+
+      it('removes bearer token from response', () => {
+        const definition: RecordingDefinition = {
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: 'Bearer secret',
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: 'secret' },
+          },
+        };
+
+        replaceCredentialInDefinition({
+          definition,
+          scheme: {
+            id: 'bearer',
+            type: SecurityType.HTTP,
+            scheme: HttpScheme.BEARER,
+          },
+          baseUrl: 'https://localhost',
+          credential: 'secret',
+        });
+
+        expect(definition).toEqual({
+          scope: 'https://localhost',
+          path: '/get?text=123',
+          method: 'GET',
+          status: 200,
+          reqheaders: {
+            ['Authorization']: `Bearer ${HIDDEN_CREDENTIALS_PLACEHOLDER}`,
+          },
+          response: {
+            some: 'data',
+            auth: { my_api_key: HIDDEN_CREDENTIALS_PLACEHOLDER },
           },
         });
       });
@@ -406,6 +630,62 @@ describe('nock utils', () => {
           path: `/get?text=123`,
           method: 'GET',
           status: 200,
+        });
+      });
+
+      it('removes parameter from raw headers', () => {
+        const parameterValue = 'integration-parameter';
+        const definition: RecordingDefinition = {
+          scope: baseUrl,
+          path: `/get?text=123`,
+          method: 'GET',
+          status: 200,
+          rawHeaders: ['Authorization', parameterValue],
+        };
+
+        replaceParameterInDefinition({
+          definition,
+          baseUrl,
+          credential: parameterValue,
+        });
+
+        expect(definition).toEqual({
+          scope: baseUrl,
+          path: `/get?text=123`,
+          method: 'GET',
+          status: 200,
+          rawHeaders: ['Authorization', HIDDEN_PARAMETERS_PLACEHOLDER],
+        });
+      });
+
+      it('removes parameter from response', () => {
+        const parameterValue = 'integration-parameter';
+        const definition: RecordingDefinition = {
+          scope: baseUrl,
+          path: `/get?text=123`,
+          method: 'GET',
+          status: 200,
+          response: {
+            some: 'data',
+            auth: { my_api_key: parameterValue },
+          },
+        };
+
+        replaceParameterInDefinition({
+          definition,
+          baseUrl,
+          credential: parameterValue,
+        });
+
+        expect(definition).toEqual({
+          scope: baseUrl,
+          path: `/get?text=123`,
+          method: 'GET',
+          status: 200,
+          response: {
+            some: 'data',
+            auth: { my_api_key: HIDDEN_PARAMETERS_PLACEHOLDER },
+          },
         });
       });
     });